refactor: move standard fields into sectionTypes/
