+++ src/schema/v2/homeView/HomeViewGenericSectionInterface.ts
@@ -1,6 +1,6 @@
 import { GraphQLInterfaceType } from "graphql"
 import { HomeViewSectionTypeNames } from "./sectionTypes/names"
-import { standardSectionFields } from "./standardSectionFields"
+import { standardSectionFields } from "./sectionTypes/standardSectionFields"
 
 export const HomeViewGenericSectionInterface = new GraphQLInterfaceType({
   name: HomeViewSectionTypeNames.HomeViewSectionGeneric,
