+++ src/schema/v2/homeView/sectionTypes/HomeViewActivitySectionType.ts
@@ -5,7 +5,7 @@ import { NotificationsConnection } from "../../notifications"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "../standardSectionFields"
+import { standardSectionFields } from "./standardSectionFields"
 
 export const HomeViewActivitySectionType = new GraphQLObjectType<
   any,
