+++ src/schema/v2/homeView/sectionTypes/HomeViewArticlesSectionType.ts
@@ -4,7 +4,7 @@ import { ResolverContext } from "types/graphql"
 import ArticlesConnection from "../../articlesConnection"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "../standardSectionFields"
+import { standardSectionFields } from "./standardSectionFields"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
 
 export const HomeViewArticlesSectionType = new GraphQLObjectType<
