+++ src/schema/v2/homeView/sectionTypes/HomeViewAuctionResultsSectionType.ts
@@ -5,7 +5,7 @@ import { auctionResultConnection } from "../../auction_result"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "../standardSectionFields"
+import { standardSectionFields } from "./standardSectionFields"
 
 export const HomeViewAuctionResultsSectionType = new GraphQLObjectType<
   any,
