+++ src/schema/v2/homeView/sectionTypes/HomeViewExploreBySectionType.ts
@@ -8,7 +8,7 @@ import { ResolverContext } from "types/graphql"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "../standardSectionFields"
+import { standardSectionFields } from "./standardSectionFields"
 import { ImageType } from "../../image"
 
 export const ExploreByMarketingCollectionCategory = new GraphQLObjectType<
