+++ src/schema/v2/homeView/sectionTypes/HomeViewFairsSectionType.ts
@@ -5,7 +5,7 @@ import { fairsConnection } from "../../fairs"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "../standardSectionFields"
+import { standardSectionFields } from "./standardSectionFields"
 
 export const HomeViewFairsSectionType = new GraphQLObjectType<
   any,
