+++ src/schema/v2/homeView/sectionTypes/HomeViewMarketingCollectionsSectionType.ts
@@ -6,7 +6,7 @@ import { MarketingCollectionType } from "../../marketingCollections"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "../standardSectionFields"
+import { standardSectionFields } from "./standardSectionFields"
 
 export const HomeViewMarketingCollectionsSectionType = new GraphQLObjectType<
   any,
