+++ src/schema/v2/homeView/sectionTypes/HomeViewViewingRoomsSectionType.ts
@@ -3,7 +3,7 @@ import { ResolverContext } from "types/graphql"
 import { NodeInterface } from "../../object_identification"
 import { HomeViewGenericSectionInterface } from "../HomeViewGenericSectionInterface"
 import { HomeViewSectionTypeNames } from "./names"
-import { standardSectionFields } from "../standardSectionFields"
+import { standardSectionFields } from "./standardSectionFields"
 
 export const HomeViewViewingRoomsSectionType = new GraphQLObjectType<
   any,
